requirements.txt
strands-agents
strands-agents-tools
pandas
matplotlib
plotly
hypothesis
pytest
pyyaml
fastapi
uvicorn[standard]
websockets
python-multipart
